I have no idea what went wrong
Formatting is a mess, despite the markdown being the exact same on my copy vs this one. Think the only way to fix is to not have the tables in tabs in markdown. Meaning, we have to manually move the tables into the correct tabs in gitbook directly. I'm just going to nuke it for now until we can figure out what to do going forward.